Admin history/view for non-admin views (issue #46)
Changes:

* Refactor admin.py by extracting common code between admin and non-admin view and putting it to two new modules: mixins.py (for view mixins) and compare.py (for compare objects)
* Implement a HistoryCompareDetailView CBV that can be used as a normal DetailView to both display the history and select-version-form *and* the changes of the selected versions
* Fix tests so that they run under windows and add test for CBV (by copying the tests for the SimpleModel -- probably this could be refactored a bit)
* Update README and AUTOHRS
