+++ AUTHORS
@@ -15,6 +15,8 @@ CONTRIBUTORS are and/or have been (alphabetic order):
   Github: <https://github.com/bmihelac> 
 * Pahaz, Blinov
   Github: <https://github.com/pahaz>
+* Papastefanos, Serafeim
+  Github: <https://github.com/spapas>  
 * Spencer, Samuel
   Github: <https://github.com/LegoStormtroopr>
 * C. Barrionuevo da Luz, Fabio
