+++ reversion_compare/templates/reversion-compare/compare_partial.html
@@ -1,0 +1,25 @@
+{% load i18n %}
+{% for field_diff in compare_data %}
+    <h3>{% firstof field_diff.field.verbose_name field_diff.field.related_name %}{% if field_diff.is_related and not field_diff.follow %}<sup class="follow">*</sup>{% endif %}</h3>
+    {% if field_diff.field.help_text %}<p class="help">{{ field_diff.field.help_text }}</p>{% endif %}
+    <div class="module">
+        <p>{{ field_diff.diff }}</p>
+    </div>
+{% empty %}
+    <div class="module">
+        <p><strong>{% trans "There are no differences." %}</strong></p>
+    </div>
+{% endfor %}
+
+<h4>{% trans "Edit comment:" %}</h4>
+<blockquote>{{ version2.revision.comment|default:"(no comment exists)" }}</blockquote>
+
+{% if has_unfollowed_fields %}
+<h4 class="follow">{% trans "Note:" %}</h4>
+<p class="follow">
+    {% blocktrans %}
+        Fields/entries marked with <sup class="follow">*</sup> are not under reversion control.
+        It may be that not all marked information are correct.
+    {% endblocktrans %}
+</p>
+{% endif %}
