+++ tests/templates/tests/simplemodel_detail.html
@@ -1,0 +1,17 @@
+
+<style type="text/css">
+    /* minimal style for the diffs */
+    del, ins {
+        color: #000;
+        text-decoration: none;
+    }
+    del { background-color: #ffe6e6; }
+    ins { background-color: #e6ffe6; }
+    sup.follow { color: #5555ff; }
+</style>
+
+
+{% include "reversion-compare/action_list_partial.html"  %}
+{% include "reversion-compare/compare_partial.html"  %}
+{% include "reversion-compare/compare_links_partial.html"  %}
+
