+++ tests/views.py
@@ -1,0 +1,7 @@
+
+from reversion_compare.views import HistoryCompareDetailView
+from .models import SimpleModel
+
+class SimpleModelHistoryCompareView(HistoryCompareDetailView):
+    model = SimpleModel
+
